Split otils.go to multiple files
Group by functions responsibility.
